+++ jdaviz/configs/default/plugins/__init__.py
@@ -8,6 +8,6 @@ from .model_fitting.model_fitting import *  # noqa
 from .collapse.collapse import *  # noqa
 from .line_lists.line_lists import *  # noqa
 from .metadata_viewer.metadata_viewer import *  # noqa
-from .export_plot.export_plot import *  # noqa
+from .export.export import *  # noqa
 from .plot_options.plot_options import *  # noqa
 from .markers.markers import * # noqa