+++ jdaviz/core/tests/test_template_mixin.py
@@ -110,17 +110,17 @@ def test_viewer_select(cubeviz_helper, spectrum1d_cube):
     fv = app.get_viewer("flux-viewer")
     sv = app.get_viewer("spectrum-viewer")
 
-    # export plot uses the mixin
-    p = app.get_tray_item_from_name('g-export-plot')
+    # export plugin uses the mixin
+    p = cubeviz_helper.plugins['Export']
     assert len(p.viewer.ids) == 3
     assert len(p.viewer.references) == 3
     assert len(p.viewer.labels) == 3
     assert p.viewer.selected_obj == fv
 
     # set by reference
-    p.viewer_selected = 'spectrum-viewer'
+    p.viewer = 'spectrum-viewer'
     assert p.viewer.selected_obj == sv
 
     # try setting based on id instead of reference
-    p.viewer_selected = p.viewer.ids[0]
-    assert p.viewer_selected == p.viewer.labels[0]
+    p.viewer = p.viewer.ids[0]
+    assert p.viewer.selected == p.viewer.labels[0]